Resolve issues with obfuscation, minor refactor
