+++ sahha-android/src/main/java/sdk/sahha/android/common/Session.kt
@@ -1,7 +1,5 @@
 package sdk.sahha.android.common
 
 object Session {
-    lateinit var appId: String
-    lateinit var appSecret: String
-    lateinit var externalId: String
+
 }
