+++ sahha-android/src/main/java/sdk/sahha/android/data/local/SahhaDatabase.kt
@@ -7,7 +7,7 @@ import com.sahha.android.model.SleepQueue
 import com.sahha.android.model.SleepQueueHistory
 import sdk.sahha.android.data.local.converter.Converter
 import sdk.sahha.android.data.local.dao.*
-import sdk.sahha.android.data.remote.dto.SleepDto
+import sdk.sahha.android.domain.model.dto.SleepDto
 import sdk.sahha.android.domain.model.activities.PreviousActivity
 import sdk.sahha.android.domain.model.activities.RecognisedActivity
 import sdk.sahha.android.domain.model.config.SahhaConfiguration
