+++ sahha-android/src/main/java/sdk/sahha/android/data/repository/AuthRepoImpl.kt
@@ -1,7 +1,6 @@
 package sdk.sahha.android.data.repository
 
 import android.content.SharedPreferences
-import android.util.Log
 import retrofit2.Response
 import sdk.sahha.android.common.ResponseCode
 import sdk.sahha.android.common.SahhaErrors
@@ -10,11 +9,10 @@ import sdk.sahha.android.common.TokenBearer
 import sdk.sahha.android.data.Constants.UERT
 import sdk.sahha.android.data.Constants.UET
 import sdk.sahha.android.data.remote.SahhaApi
-import sdk.sahha.android.data.remote.dto.send.ExternalIdSendDto
-import sdk.sahha.android.data.remote.dto.send.RefreshTokenSendDto
+import sdk.sahha.android.domain.model.dto.send.ExternalIdSendDto
+import sdk.sahha.android.domain.model.dto.send.RefreshTokenSendDto
 import sdk.sahha.android.domain.model.auth.TokenData
 import sdk.sahha.android.domain.repository.AuthRepo
-import sdk.sahha.android.source.SahhaConverterUtility
 
 private const val tag = "AuthRepoImpl"
 class AuthRepoImpl(
