+++ sahha-android/src/main/java/sdk/sahha/android/domain/model/device_info/DeviceInformation.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 import sdk.sahha.android.BuildConfig
 import sdk.sahha.android.data.Constants.PLATFORM_NAME
-import sdk.sahha.android.data.remote.dto.send.DeviceInformationSendDto
+import sdk.sahha.android.domain.model.dto.send.DeviceInformationSendDto
 import sdk.sahha.android.source.Sahha
 
 @Entity
