+++ sahha-android/src/main/java/sdk/sahha/android/domain/model/dto/DemographicDto.kt
@@ -1,4 +1,4 @@
-package sdk.sahha.android.data.remote.dto
+package sdk.sahha.android.domain.model.dto
 
 import androidx.annotation.Keep
 import androidx.room.Entity
