+++ sahha-android/src/main/java/sdk/sahha/android/domain/model/dto/send/RefreshTokenSendDto.kt
@@ -1,0 +1,8 @@
+package sdk.sahha.android.domain.model.dto.send
+
+import androidx.annotation.Keep
+
+@Keep
+data class RefreshTokenSendDto(
+    val refreshToken: String
+)
