+++ sahha-android/src/main/java/sdk/sahha/android/source/SahhaConverterUtility.kt
@@ -9,12 +9,12 @@ import okhttp3.ResponseBody
 import okio.Buffer
 import org.json.JSONArray
 import org.json.JSONObject
-import sdk.sahha.android.data.remote.dto.SleepDto
-import sdk.sahha.android.data.remote.dto.StepDto
-import sdk.sahha.android.data.remote.dto.send.DeviceInformationSendDto
-import sdk.sahha.android.data.remote.dto.send.PhoneUsageSendDto
-import sdk.sahha.android.data.remote.dto.send.SleepSendDto
-import sdk.sahha.android.data.remote.dto.toSleepSendDto
+import sdk.sahha.android.domain.model.dto.SleepDto
+import sdk.sahha.android.domain.model.dto.StepDto
+import sdk.sahha.android.domain.model.dto.send.DeviceInformationSendDto
+import sdk.sahha.android.domain.model.dto.send.PhoneUsageSendDto
+import sdk.sahha.android.domain.model.dto.send.SleepSendDto
+import sdk.sahha.android.domain.model.dto.toSleepSendDto
 import sdk.sahha.android.domain.model.device.PhoneUsage
 import sdk.sahha.android.domain.model.device.toPhoneUsageSendDto
 import sdk.sahha.android.domain.model.device_info.DeviceInformation
